test(CrewAttendanceManagerTest): 출석 확인 기능 테스트 추가
